solved

+++ P1/Ejercicio13/monitor.c
@@ -9,10 +9,11 @@
 #include "monitor.h"
 
 int monitor(int monitorPipe, int minerPipe){
-    int datos[2];
+    long datos[2];
     char respuesta;
     ssize_t nbytes;
-    while((nbytes = read(minerPipe, &datos, sizeof(int)*2))){
+    while((nbytes = read(minerPipe, &datos, sizeof(long)*2))){
+    // printf("%ld , %ld\n", datos[0], datos[1]);
         if (pow_hash(datos[1]) == datos[0]){
             respuesta = 1;
             printf("Solution accepted : %08d --> %08d\n", datos[0], datos[1]);
@@ -23,6 +24,7 @@ int monitor(int monitorPipe, int minerPipe){
         }
         write(monitorPipe, &respuesta, sizeof(char));
     }
+    
     if (nbytes < 0)
         exit(EXIT_FAILURE);
     exit(EXIT_SUCCESS);
